Move PolledTimeout to own libraries for source code compatibility with ESP8266
