+++ libraries/PolledTimeout/library.properties
@@ -1,0 +1,9 @@
+name=PolledTimeout
+version=1.0
+author=
+maintainer=
+sentence=
+paragraph=
+category=Other
+url=
+architectures=esp32
